requirements.txt
streamlit
langchain
langchain_groq  
langchain_openai
beautifulsoup4
langchain_community
python-dotenv
chromadb